[add] popup showTips
