+++ src/pages/Control/Control.css
@@ -32,11 +32,11 @@ html, body {
   margin-top: var(--interval);
 }
 
-.filtered-floor-tips.show {
+.bottom-tips.show {
   opacity: 1;
   bottom: 0px;
 }
-.filtered-floor-tips {
+.bottom-tips {
   transition: opacity 382ms, bottom 382ms;
   opacity: 0;
   position: absolute;
