+++ src/pages/Popup/floors/1-AddToPoker.css
@@ -5,13 +5,6 @@
   line-height: var(--footer-height);
 
   text-align: center;
-  border-top: solid 1px hsl(240deg 2% 23% / 15%);
-
-  box-shadow:
-    inset 0 var(--footer-shadow-size) var(--footer-shadow-size) calc(-1 * var(--footer-shadow-size)) var(--footer-shadow-color),
-    inset 0 calc(-1 * var(--footer-shadow-size)) var(--footer-shadow-size) calc(-1 * var(--footer-shadow-size)) var(--footer-shadow-color)
-  ;
-  background-color: var(--footer-background);
 
   height: var(--popup-height);
